refactor(mobile): extract shared auth validation/error helpers and harden disabled submit handling in auth form
