+++ apps/mobile/src/app/screens/LoginScreen.tsx
@@ -2,35 +2,7 @@ import React, { useState } from 'react';
 import { signInWithEmailPassword } from '@abstrack/supabase';
 import { getMobileSupabaseClient } from '../../lib/supabase-wiring';
 import { AuthForm } from '../components/AuthForm';
-
-function validateLoginInput(email: string, password: string): string | null {
-  const trimmedEmail = email.trim();
-  const hasEmailFormat = /.+@.+\..+/.test(trimmedEmail);
-
-  if (!trimmedEmail || !password) {
-    return 'Enter your email and password.';
-  }
-
-  if (!hasEmailFormat) {
-    return 'Enter a valid email address.';
-  }
-
-  return null;
-}
-
-function mapAuthError(message: string): string {
-  const normalized = message.toLowerCase();
-
-  if (normalized.includes('invalid login credentials')) {
-    return 'Email or password is incorrect.';
-  }
-
-  if (normalized.includes('email not confirmed')) {
-    return 'Please confirm your email before signing in.';
-  }
-
-  return message;
-}
+import { mapAuthError, validateEmailPassword } from '../auth-helpers';
 
 export function LoginScreen({ onGoToSignup }: { onGoToSignup: () => void }) {
   const [email, setEmail] = useState('');
@@ -39,7 +11,7 @@ export function LoginScreen({ onGoToSignup }: { onGoToSignup: () => void }) {
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
 
   const handleLogin = async () => {
-    const validationError = validateLoginInput(email, password);
+    const validationError = validateEmailPassword(email, password);
     if (validationError) {
       setErrorMessage(validationError);
       return;
